requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.5.2=pypi_0
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotli-python=1.0.9=py310h1105856_7
brotlipy=0.7.0=py310hf8d0d8f_1004
bzip2=1.0.8=h3422bc3_4
ca-certificates=2022.9.14=h4653dfc_0
certifi=2022.9.14=pyhd8ed1ab_0
cffi=1.15.1=py310he00a5c5_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.1.3=py310hbe9552e_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py310h834c97f_0
cycler=0.11.0=pyhd8ed1ab_0
dash=2.6.1=pyhd8ed1ab_0
dash-bootstrap-components=1.2.1=pyhd8ed1ab_0
dash-table=5.0.0=pyhd8ed1ab_1
django=4.1.1=pypi_0
elasticsearch=7.9.1=pyh9f0ad1d_0
flask=2.1.2=pyhd8ed1ab_1
flask-compress=1.12=pyhd8ed1ab_1
flask-paginate=2022.1.8=pyhd8ed1ab_0
fonttools=4.34.4=py310h02f21da_0
freetype=2.10.4=h17b34a0_1
frozendict=2.3.0=pypi_0
giflib=5.2.1=h27ca646_2
idna=3.3=pyhd8ed1ab_0
importlib-metadata=4.11.4=py310hbe9552e_0
itsdangerous=2.1.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
jq=1.6=h27ca646_1000
kiwisolver=1.4.4=py310hd23d0e8_0
lcms2=2.12=had6a04f_0
lerc=3.0=hbdafb3b_0
libblas=3.9.0=15_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=15_osxarm64_openblas
libcxx=14.0.6=h04bba0f_0
libdeflate=1.12=he4db4b2_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
liblapack=3.9.0=15_osxarm64_openblas
libopenblas=0.3.20=openmp_h2209c59_0
libpng=1.6.37=h2c9beb0_3
libtiff=4.4.0=hcbbed22_1
libwebp=1.2.2=h0d20362_0
libwebp-base=1.2.2=h3422bc3_1
libxcb=1.13=h9b22ae9_1004
libzlib=1.2.12=ha287fd2_2
llvm-openmp=14.0.4=hd125106_0
lz4-c=1.9.3=hbdafb3b_1
markupsafe=2.1.1=py310hf8d0d8f_1
matplotlib=3.5.2=py310hb6292c7_0
matplotlib-base=3.5.2=py310hbeb1b0d_0
matplotlib-venn=0.11.7=pyhd8ed1ab_0
munkres=1.0.7=py_1
ncurses=6.3=h07bb92c_1
nltk=3.6.7=pyhd8ed1ab_0
numpy=1.23.1=py310h0a343b5_0
oniguruma=6.9.8=he4db4b2_0
openjpeg=2.4.0=h062765e_1
openssl=3.0.5=h03a7124_2
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=py310ha6a5cd6_2
pigz=2.6=h1a28f6b_0
pillow=9.2.0=py310hc9df86f_0
pip=22.1.2=pyhd8ed1ab_0
plotly=5.10.0=pyhd8ed1ab_0
pthread-stubs=0.4=h27ca646_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310hbe9552e_5
python=3.10.5=h4eee789_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.1=pyhd8ed1ab_0
readline=8.1.2=h46ed386_0
regex=2022.7.9=py310h02f21da_0
requests=2.28.1=pyhd8ed1ab_0
scipy=1.8.1=py310hdb41229_0
setuptools=63.2.0=py310hbe9552e_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.1=h40dfcc0_0
sqlparse=0.4.2=pypi_0
sugo=0.1.20220825=dev_0
tenacity=8.0.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tornado=6.2=py310h02f21da_0
tqdm=4.64.0=pyhd8ed1ab_0
tzdata=2022a=h191b570_0
unicodedata2=14.0.0=py310hf8d0d8f_1
urllib3=1.26.10=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.37.1=pyhd8ed1ab_0
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.5=h642e427_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.12=ha287fd2_2
zstd=1.5.2=hd705a24_2
